refactor: remove axios; serve Horizon stats via an Inertia prop (B2)
axios was down to two usages after B3 + the InfiniteScroll rollout. Remove both
and drop the JSON-polling endpoint the HorizonStats widget used:

- New QueueStatsService holds the queue_count/error_count/status computation
  (moved verbatim from the deleted QueueController).
- HandleInertiaRequests exposes it as an `Inertia::optional('queueStats')` shared
  prop — resolved only when requested (never on ordinary loads), mirroring the
  existing affiliatedEntities pattern.
- HorizonStats reads the `queueStats` prop and refreshes it every 5s via an
  Inertia partial reload (router.reload with only:['queueStats']); the axios
  poll of /queue/status is gone.
- Deleted the now-unused horizon.status route + QueueController.
- bootstrap.js: removed the dead `window.axios` global (nothing referenced it;
  Functions/http.js already sends X-Requested-With for non-Inertia JSON calls).

axios is not a declared dependency, so there is no package.json change. No real
axios usage remains in resources/js.

Verified: Pint, PHPStan, type-coverage 100%, ESLint 0 errors; render/feature
suites pass. HorizonStats rendering + polling validated by the core Browser CI.
